hotfix: scoring 필드 Integer → Short (SMALLINT 매핑 수정) (#84)
- Flyway V6에서 SMALLINT로 생성한 컬럼과 JPA 타입 불일치 수정
- TrendItem: Integer → Short
- AnalysisService: setter 호출 시 (short) 캐스트
- TrendItemResponse: Map<String, Integer> → Map<String, Short>

+++ mud-backend/src/main/java/com/mud/domain/entity/TrendItem.java
@@ -91,16 +91,16 @@ public class TrendItem {
     private String arxivId;
 
     @Column(name = "scoring_relevance")
-    private Integer scoringRelevance;
+    private Short scoringRelevance;
 
     @Column(name = "scoring_timeliness")
-    private Integer scoringTimeliness;
+    private Short scoringTimeliness;
 
     @Column(name = "scoring_actionability")
-    private Integer scoringActionability;
+    private Short scoringActionability;
 
     @Column(name = "scoring_impact")
-    private Integer scoringImpact;
+    private Short scoringImpact;
 
     @Column(name = "topic_tag", length = 100)
     private String topicTag;
